arm: DT: msm8939-common: Update temperature sensor support
Update sensor id mapping of temperature sensors (TSENS) supported
for MSM8939 target. Remove support to read from TSENS sensor 4.

Add qcom,valid-status-check property to check the valid bit in the status
register when reading the ADC data. If the valid bit is not set proceed to
read the data register again.
Thermal on MSM8939/36 requires qcom,valid-status-check. MSM8939/36 is a
chip that gets very hot and we can not allow us to stay without thermal
protection online for even one boot if we want to keep our boards safe
and alive.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>